LayoutTests/platform/efl/fast/regions/webkit-named-flow-event-to-null-expected.txt

Tests regionLayoutEvent being dispatched when the content is removed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Flow content
PASS event.target.name is "article"
Removing region_2 from flow
Removing region_1 from flow
Removing article from flow
PASS layoutUpdateDispatchedForNullFlow is false
PASS successfullyParsed is true

TEST COMPLETE
